refactor : PR 리뷰 후 수정
