+++ app/api/common-api/src/main/java/org/example/filter/JWTFilter.java
@@ -8,12 +8,15 @@ import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.example.exception.BusinessException;
+import org.example.repository.TokenRepository;
 import org.example.security.dto.AuthenticatedUser;
 import org.example.security.dto.TokenParam;
 import org.example.security.dto.UserParam;
 import org.example.security.token.JWTGenerator;
 import org.example.security.token.JWTHandler;
 import org.example.security.token.RefreshTokenProcessor;
+import org.example.security.vo.TokenError;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -27,6 +30,7 @@ public class JWTFilter extends OncePerRequestFilter {
     private final JWTHandler jwtHandler;
     private final JWTGenerator jwtGenerator;
     private final RefreshTokenProcessor refreshTokenProcessor;
+    private final TokenRepository tokenRepository;
 
     @Override
     protected void doFilterInternal(
@@ -50,10 +54,14 @@ public class JWTFilter extends OncePerRequestFilter {
     private void handleAccessToken(HttpServletRequest request) {
         String accessToken = jwtHandler.extractAccessToken(request);
         UserParam userParam = jwtHandler.extractUserFrom(accessToken);
-
-        jwtGenerator.verifyLogoutAccessToken(userParam);
-
+        verifyLogoutAccessToken(userParam);
         saveOnSecurityContextHolder(userParam);
+    }
+
+    public void verifyLogoutAccessToken(UserParam userParam) {
+        if (tokenRepository.existAccessToken(userParam.userId().toString())) {
+            throw new BusinessException(TokenError.INVALID_TOKEN);
+        }
     }
 
     private void saveOnSecurityContextHolder(UserParam userParam) {
