+++ app/api/common-api/src/main/java/org/example/repository/TokenRepository.java
@@ -8,7 +8,7 @@ public interface TokenRepository {
 
     void save(String userId, String refreshToken);
 
-    Optional<String> getOldRefreshToken(String userId);
+    Optional<String> getExistRefreshToken(String userId);
 
     Boolean existAccessToken(String userId);
 }
