+++ app/api/common-api/src/main/java/org/example/vo/UserRoleApiType.java
@@ -1,6 +1,7 @@
 package org.example.vo;
 
 import lombok.Getter;
+import org.example.entity.User;
 
 @Getter
 public enum UserRoleApiType {
@@ -12,5 +13,9 @@ public enum UserRoleApiType {
 
     UserRoleApiType(String authority) {
         this.authority = authority;
+    }
+
+    public static UserRoleApiType from(User user) {
+        return UserRoleApiType.valueOf(user.getUserRole().name());
     }
 }
