+++ app/api/user-api/src/main/java/org/example/service/dto/request/LoginServiceRequest.java
@@ -1,16 +1,16 @@
 package org.example.service.dto.request;
 
 import lombok.Builder;
+import org.example.entity.User;
 import org.example.entity.credential.SocialCredentials;
-import org.example.service.dto.LoginServiceDto;
 
 @Builder
 public record LoginServiceRequest(
     SocialCredentials socialCredentials
 ) {
 
-    public LoginServiceDto toLoginServiceDto() {
-        return LoginServiceDto.builder()
+    public User toUser() {
+        return User.builder()
             .socialCredentials(socialCredentials)
             .build();
     }
