+++ app/domain/user-domain/src/main/java/org/example/vo/UserGender.java
@@ -1,5 +1,5 @@
 package org.example.vo;
 
 public enum UserGender {
-    MAN, WOMAN, NONE
+    MAN, WOMAN, NOT_CHOSEN
 }
